# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for the SweetRPG Game Room domain: a user's library (owned catalog volumes), wishlists, play tables, and peer-to-peer volume loans, each with per-resource visibility controls.",
        "title": "Game Room API service",
        "termsOfService": "https://pilgrimagesoftware.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://sweetrpg.com",
            "email": "admin@sweetrpg.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://mit-license.org/"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/users/{user_id}/wishlists": {
            "get": {
                "description": "List a user's wishlists, filtered to what the caller may see.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "List wishlists",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user-123",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/server.WishlistVO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new named wishlist, defaulting to private visibility.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "Create wishlist",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user-123",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Wishlist name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.createWishlistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.WishlistVO"
                        }
                    },
                    "400": {
                        "description": "name missing",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/wishlists/{wishlist_id}": {
            "get": {
                "description": "Get one wishlist by ID, if the caller may see it.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "Get wishlist",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user-123",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "wl-789",
                        "description": "Wishlist ID",
                        "name": "wishlist_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.WishlistVO"
                        }
                    },
                    "404": {
                        "description": "no wishlist with this ID exists, or the caller may not see it",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    }
                }
            },
            "delete": {
                "description": "Permanently remove a wishlist. Owner-only.",
                "tags": [
                    "wishlist"
                ],
                "summary": "Delete wishlist",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user-123",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "wl-789",
                        "description": "Wishlist ID",
                        "name": "wishlist_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "wishlist exists but is owned by a different user",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "404": {
                        "description": "no wishlist with this ID exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/wishlists/{wishlist_id}/entries": {
            "post": {
                "description": "Add a catalog volume to a wishlist. Owner-only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "Add wishlist entry",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user-123",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "wl-789",
                        "description": "Wishlist ID",
                        "name": "wishlist_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Volume to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.volumeEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.WishlistVO"
                        }
                    },
                    "400": {
                        "description": "volume_id missing",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "403": {
                        "description": "wishlist exists but is owned by a different user",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "404": {
                        "description": "no wishlist with this ID exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/wishlists/{wishlist_id}/entries/{volume_id}": {
            "delete": {
                "description": "Remove a catalog volume from a wishlist. Owner-only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "Remove wishlist entry",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user-123",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "wl-789",
                        "description": "Wishlist ID",
                        "name": "wishlist_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "vol-123",
                        "description": "Volume ID",
                        "name": "volume_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.WishlistVO"
                        }
                    },
                    "403": {
                        "description": "wishlist exists but is owned by a different user",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "404": {
                        "description": "no wishlist with this ID exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/wishlists/{wishlist_id}/visibility": {
            "put": {
                "description": "Set a wishlist's visibility. Owner-only. Valid visibility values: public, friends, friends_of_friends, private.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "Set wishlist visibility",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user-123",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "wl-789",
                        "description": "Wishlist ID",
                        "name": "wishlist_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New visibility",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.visibilityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.WishlistVO"
                        }
                    },
                    "400": {
                        "description": "invalid body or invalid visibility value",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "403": {
                        "description": "wishlist exists but is owned by a different user",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "404": {
                        "description": "no wishlist with this ID exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "server.WishlistVO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "deleted_by": {
                    "type": "string"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.WishlistEntryVO"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "visibility": {
                    "type": "string"
                }
            }
        },
...
        "server.createWishlistRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Birthday wishlist"
                }
            }
        },
...
        "server.visibilityRequest": {
            "type": "object",
            "properties": {
                "visibility": {
                    "type": "string",
                    "example": "friends"
                }
            }
        },
        "server.volumeEntryRequest": {
            "type": "object",
            "properties": {
                "volume_id": {
                    "type": "string",
                    "example": "vol-123"
                },
                "volume_title": {
                    "type": "string",
                    "example": "Curse of Strahd"
                }
            }
        },
        "vo.ErrorVO": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "vo.WishlistEntryVO": {
            "type": "object",
            "properties": {
                "added_at": {
                    "type": "string"
                },
                "volume_id": {
                    "type": "string"
                },
                "volume_title": {
                    "type": "string"
                }
            }
        }
    }
}